Changes for readr 2.0.0
In readr 2.0.0 blank lines are not implicitly skipped. It seems the
GloBI service returns data with a trailing blank line, so we need to
drop these rows with all NA values.

In addition testthat supports a helper file which is automatically
sourced before running your tests or when using `devtools::load_all()`.
I moved the test utils functions into this file.

I also suppressed the parsing warnings as well as the messages, as that
clears up some spurious output when running the tests.